refactor: Split the unit testing for timeout_check and remove_nsfw_sites into 2 separate files
